nax: deduplicate NAX class hierarchy into shared nax_common.h
Both mlx/backend/metal/kernels/steel/gemm/nax.h and
mlx/backend/metal/kernels/steel/attn/nax.h now re-export from a
single nax_common.h containing the canonical Role enum,
BaseNAXFrag, NAXFrag32, NAXTile, and tile_matmad_nax. The attn/
tree had drifted ~900 lines behind gemm/'s Phase 1/2/5/6
additions (no NAXFrag32, no Role-aware load); dedup brings them
back in sync.

Added three kPacking==2-only convenience overloads to NAXTile
(load(src,ld), load_rows(src,ld,n_rows), store(dst,ld)) that
match the old attn/nax.h API so steel_attention_nax.h compiles
without changes; static_asserts guard against kPacking==1 misuse.

Phase 7 builds on this dedup to enable NAX SDPA on g16.
