+++ mlx/backend/metal/kernels/steel/attn/nax.h
@@ -1,887 +1,7 @@
-// Copyright © 2025 Apple Inc.
+// Copyright © 2024-26 Apple Inc.
 
 #pragma once
 
-#include <metal_simdgroup>
-#include <metal_simdgroup_matrix>
-#include <metal_stdlib>
+#include "mlx/backend/metal/kernels/steel/nax_common.h"
 
-#include "mlx/backend/metal/kernels/steel/defines.h"
-#include "mlx/backend/metal/kernels/steel/utils/integral_constant.h"
-
-#include <MetalPerformancePrimitives/MetalPerformancePrimitives.h>
-
-using namespace metal;
-
-///////////////////////////////////////////////////////////////////////////////
-// MMA helper
-///////////////////////////////////////////////////////////////////////////////
-
-namespace mlx {
-namespace steel {
-
-///////////////////////////////////////////////////////////////////////////////
-// NAX Steel with new tiles
-///////////////////////////////////////////////////////////////////////////////
-
-struct BaseNAXFrag {
-  STEEL_CONST short kFragRows = 16;
-  STEEL_CONST short kFragCols = 16;
-
-  STEEL_CONST short kElemsPerFrag = (kFragRows * kFragCols) / 32;
-
-  STEEL_CONST short kElemRows = 2;
-  STEEL_CONST short kElemCols = 4;
-
-  STEEL_CONST short kElemRowsJump = 8;
-
-  static_assert(
-      kElemRows * kElemCols == kElemsPerFrag,
-      "MMAFrag shape is not consistent with MMAFrag size");
-
-  template <typename U>
-  using dtype_frag_t = typename metal::vec<U, kElemsPerFrag>;
-
-  METAL_FUNC static short2 get_coord() {
-    const ushort simd_lane_id = __metal_get_thread_index_in_simdgroup(ushort());
-    const short qid = simd_lane_id >> 2;
-    const short fm = ((qid & 4) | ((simd_lane_id >> 1) & 3));
-    const short fn = ((qid & 2) | (simd_lane_id & 1)) * 4;
-    return short2{fn, fm};
-  }
-
-  METAL_FUNC static short2 get_coord(short idx) {
-    const ushort simd_lane_id = __metal_get_thread_index_in_simdgroup(ushort());
-    const short qid = simd_lane_id >> 2;
-    const short fm = ((qid & 4) | ((simd_lane_id >> 1) & 3)) + (idx >> 2) * 8;
-    const short fn = ((qid & 2) | (simd_lane_id & 1)) * 4 + idx % 4;
-    return short2{fn, fm};
-  }
-
-  template <
-      typename T,
-      typename SrcPtrType,
-      typename StrX,
-      typename StrY,
-      typename OffX = Int<0>,
-      typename OffY = Int<0>>
-  METAL_FUNC static constexpr void load(
-      thread dtype_frag_t<T>& dst,
-      SrcPtrType src,
-      StrX str_x,
-      StrY str_y,
-      OffX off_x = {},
-      OffY off_y = {}) {
-    const short2 sc = get_coord();
-    src += sc.y * str_x + sc.x * str_y;
-
-    STEEL_PRAGMA_UNROLL
-    for (short i = 0; i < kElemRows; i++) {
-      const auto r = off_x + i * kElemRowsJump;
-      const auto c = off_y;
-
-      if constexpr (metal::is_same_v<StrY, Int<1>>) {
-        STEEL_PRAGMA_UNROLL
-        for (short j = 0; j < kElemCols; j++) {
-          dst[i * kElemCols + j] = static_cast<T>(src[r * str_x + c + j]);
-        }
-      } else {
-        STEEL_PRAGMA_UNROLL
-        for (short j = 0; j < kElemCols; j++) {
-          dst[i * kElemCols + j] =
-              static_cast<T>(src[r * str_x + (c + j) * str_y]);
-        }
-      }
-    }
-  }
-
-  template <
-      typename T,
-      typename SrcPtrType,
-      typename StrX,
-      typename StrY,
-      typename LimX,
-      typename OffX = Int<0>,
-      typename OffY = Int<0>>
-  METAL_FUNC static constexpr void load_rows(
-      thread dtype_frag_t<T>& dst,
-      SrcPtrType src,
-      StrX str_x,
-      StrY str_y,
-      LimX lim_x,
-      OffX off_x = {},
-      OffY off_y = {}) {
-    const short2 sc = get_coord();
-    src += sc.y * str_x + sc.x * str_y;
-    auto lx = lim_x - sc.y;
-
-    STEEL_PRAGMA_UNROLL
-    for (short i = 0; i < kElemRows; i++) {
-      const auto r = off_x + i * kElemRowsJump;
-      const auto c = off_y;
-
-      if (r < lx) {
-        if constexpr (metal::is_same_v<StrY, Int<1>>) {
-          STEEL_PRAGMA_UNROLL
-          for (short j = 0; j < kElemCols; j++) {
-            dst[i * kElemCols + j] = static_cast<T>(src[r * str_x + (c + j)]);
-          }
-        } else {
-          STEEL_PRAGMA_UNROLL
-          for (short j = 0; j < kElemCols; j++) {
-            dst[i * kElemCols + j] =
-                static_cast<T>(src[r * str_x + (c + j) * str_y]);
-          }
-        }
-
-      } else {
-        STEEL_PRAGMA_UNROLL
-        for (short j = 0; j < kElemCols; j++) {
-          dst[i * kElemCols + j] = T(0);
-        }
-      }
-    }
-  }
-
-  template <
-      typename T,
-      typename SrcPtrType,
-      typename StrX,
-      typename StrY,
-      typename LimX,
-      typename LimY,
-      typename OffX = Int<0>,
-      typename OffY = Int<0>>
-  METAL_FUNC static constexpr void load_safe(
-      thread dtype_frag_t<T>& dst,
-      SrcPtrType src,
-      StrX str_x,
-      StrY str_y,
-      LimX lim_x,
-      LimY lim_y,
-      OffX off_x = {},
-      OffY off_y = {}) {
-    const short2 sc = get_coord();
-    src += sc.y * str_x + sc.x * str_y;
-    auto lx = lim_x - sc.y;
-    auto ly = lim_y - sc.x;
-
-    STEEL_PRAGMA_UNROLL
-    for (short i = 0; i < kElemRows; i++) {
-      const auto r = off_x + i * kElemRowsJump;
-      const auto c = off_y;
-      STEEL_PRAGMA_UNROLL
-      for (short j = 0; j < kElemCols; j++) {
-        if ((r < lx) && ((c + j) < ly)) {
-          dst[i * kElemCols + j] =
-              static_cast<T>(src[r * str_x + (c + j) * str_y]);
-        } else {
-          dst[i * kElemCols + j] = T(0);
-        }
-      }
-    }
-  }
-
-  template <
-      typename T,
-      typename DstPtrType,
-      typename StrX,
-      typename StrY,
-      typename OffX = Int<0>,
-      typename OffY = Int<0>>
-  METAL_FUNC static constexpr void store(
-      const thread dtype_frag_t<T>& src,
-      DstPtrType dst,
-      StrX str_x,
-      StrY str_y,
-      OffX off_x = {},
-      OffY off_y = {}) {
-    using U = pointer_element_t<DstPtrType>;
-
-    const short2 sc = get_coord();
-    dst += sc.y * str_x + sc.x * str_y;
-
-    STEEL_PRAGMA_UNROLL
-    for (short i = 0; i < kElemRows; i++) {
-      const auto r = off_x + i * kElemRowsJump;
-      const auto c = off_y;
-
-      if constexpr (metal::is_same_v<StrY, Int<1>>) {
-        STEEL_PRAGMA_UNROLL
-        for (short j = 0; j < kElemCols; j++) {
-          dst[r * str_x + c + j] = static_cast<U>(src[i * kElemCols + j]);
-        }
-      } else {
-        STEEL_PRAGMA_UNROLL
-        for (short j = 0; j < kElemCols; j++) {
-          dst[r * str_x + (c + j) * str_y] =
-              static_cast<U>(src[i * kElemCols + j]);
-        }
-      }
-    }
-  }
-
-  template <
-      typename T,
-      typename DstPtrType,
-      typename StrX,
-      typename StrY,
-      typename LimX,
-      typename OffX = Int<0>,
-      typename OffY = Int<0>>
-  METAL_FUNC static constexpr void store_rows(
-      const thread dtype_frag_t<T>& src,
-      DstPtrType dst,
-      StrX str_x,
-      StrY str_y,
-      LimX lim_x,
-      OffX off_x = {},
-      OffY off_y = {}) {
-    using U = pointer_element_t<DstPtrType>;
-
-    const short2 sc = get_coord();
-    dst += sc.y * str_x + sc.x * str_y;
-    auto lx = lim_x - sc.y;
-
-    STEEL_PRAGMA_UNROLL
-    for (short i = 0; i < kElemRows; i++) {
-      const auto r = off_x + i * kElemRowsJump;
-      const auto c = off_y;
-
-      if (r < lx) {
-        if constexpr (metal::is_same_v<StrY, Int<1>>) {
-          STEEL_PRAGMA_UNROLL
-          for (short j = 0; j < kElemCols; j++) {
-            dst[r * str_x + c + j] = static_cast<U>(src[i * kElemCols + j]);
-          }
-        } else {
-          STEEL_PRAGMA_UNROLL
-          for (short j = 0; j < kElemCols; j++) {
-            dst[r * str_x + (c + j) * str_y] =
-                static_cast<U>(src[i * kElemCols + j]);
-          }
-        }
-      }
-    }
-  }
-
-  template <
-      typename T,
-      typename DstPtrType,
-      typename StrX,
-      typename StrY,
-      typename LimX,
-      typename LimY,
-      typename OffX = Int<0>,
-      typename OffY = Int<0>>
-  METAL_FUNC static constexpr void store_safe(
-      const thread dtype_frag_t<T>& src,
-      DstPtrType dst,
-      StrX str_x,
-      StrY str_y,
-      LimX lim_x,
-      LimY lim_y,
-      OffX off_x = {},
-      OffY off_y = {}) {
-    using U = pointer_element_t<DstPtrType>;
-
-    const short2 sc = get_coord();
-    dst += sc.y * str_x + sc.x * str_y;
-    auto lx = lim_x - sc.y;
-    auto ly = lim_y - sc.x;
-
-    STEEL_PRAGMA_UNROLL
-    for (short i = 0; i < kElemRows; i++) {
-      const auto r = off_x + i * kElemRowsJump;
-      const auto c = off_y;
-
-      STEEL_PRAGMA_UNROLL
-      for (short j = 0; j < kElemCols; j++) {
-        if (r < lx && (c + j) < ly) {
-          dst[r * str_x + (c + j) * str_y] =
-              static_cast<U>(src[i * kElemCols + j]);
-        }
-      }
-    }
-  }
-
-  template <
-      typename T,
-      typename DstPtrType,
-      typename StrX,
-      typename StrY,
-      typename StartX,
-      typename StopX,
-      typename StartY,
-      typename StopY,
-      typename OffX = Int<0>,
-      typename OffY = Int<0>>
-  METAL_FUNC static constexpr void store_slice(
-      const thread dtype_frag_t<T>& src,
-      DstPtrType dst,
-      StrX str_x,
-      StrY str_y,
-      StartX start_x,
-      StopX stop_x,
-      StartY start_y,
-      StopY stop_y,
-      OffX off_x = Int<0>{},
-      OffY off_y = Int<0>{}) {
-    using U = pointer_element_t<DstPtrType>;
-
-    const short2 sc = get_coord();
-
-    const_for_loop<0, kElemRows, 1>([&](auto idx_row) {
-      const auto r = off_x + idx_row * Int<kElemRowsJump>{};
-      if (r >= stop_x - sc.y || r < start_x - sc.y) {
-        return;
-      }
-
-      const_for_loop<0, kElemCols, 1>([&](auto idx_col) {
-        const auto c = off_y + idx_col;
-        if (c >= stop_y - sc.x || c < start_y - sc.x) {
-          return;
-        }
-
-        const auto src_idx = idx_row * Int<kElemCols>{} + idx_col;
-        dst[(r + sc.y) * str_x + (c + sc.x) * str_y] =
-            static_cast<U>(src[src_idx]);
-      });
-    });
-  }
-
-  template <typename Op, typename T>
-  METAL_FUNC static constexpr void row_reduce(
-      thread const dtype_frag_t<T>& inp_vals,
-      thread T* reduced_vals) {
-    STEEL_PRAGMA_UNROLL
-    for (short i = 0; i < kElemRows; i++) {
-      T thr_reduce = Op::apply(
-          Op::apply(inp_vals[i * kElemCols + 0], inp_vals[i * kElemCols + 1]),
-          Op::apply(inp_vals[i * kElemCols + 2], inp_vals[i * kElemCols + 3]));
-
-      T qgr_reduce = simd_shuffle_xor(thr_reduce, ushort(1));
-      qgr_reduce = Op::apply(thr_reduce, qgr_reduce);
-
-      T sgr_reduce = simd_shuffle_xor(qgr_reduce, ushort(8));
-      sgr_reduce = Op::apply(qgr_reduce, sgr_reduce);
-
-      reduced_vals[i] = Op::apply(reduced_vals[i], sgr_reduce);
-    }
-  }
-
-  template <typename Op, typename T>
-  METAL_FUNC static constexpr void row_bin_op(
-      thread dtype_frag_t<T>& inp_vals,
-      thread T* row_vals) {
-    STEEL_PRAGMA_UNROLL
-    for (short i = 0; i < kElemRows; i++) {
-      STEEL_PRAGMA_UNROLL
-      for (short j = 0; j < kElemCols; j++) {
-        inp_vals[i * kElemCols + j] =
-            Op::apply(inp_vals[i * kElemCols + j], row_vals[i]);
-      }
-    }
-  }
-
-  template <
-      typename CType,
-      typename AType,
-      typename BType,
-      bool transpose_a = false,
-      bool transpose_b = false>
-  METAL_FUNC static constexpr void mma(
-      thread dtype_frag_t<CType>& Cn0,
-      thread dtype_frag_t<CType>& Cn1,
-      const thread dtype_frag_t<AType>& A,
-      metal::bool_constant<transpose_a>,
-      const thread dtype_frag_t<BType>& Bn0,
-      const thread dtype_frag_t<BType>& Bn1,
-      metal::bool_constant<transpose_b>) {
-    constexpr auto desc = mpp::tensor_ops::matmul2d_descriptor(
-        16,
-        32,
-        16,
-        transpose_a,
-        transpose_b,
-        true,
-        mpp::tensor_ops::matmul2d_descriptor::mode::multiply_accumulate);
-
-    // Create matmul op
-    mpp::tensor_ops::matmul2d<desc, metal::execution_simdgroup> gemm_op;
-
-    // Create matmul operands in registers
-    auto ct_a =
-        gemm_op
-            .template get_left_input_cooperative_tensor<AType, BType, CType>();
-    auto ct_b =
-        gemm_op
-            .template get_right_input_cooperative_tensor<AType, BType, CType>();
-
-    // Create matmul output in register
-    auto ct_c = gemm_op.template get_destination_cooperative_tensor<
-        decltype(ct_a),
-        decltype(ct_b),
-        CType>();
-
-    // Load A in to left operand registers
-    STEEL_PRAGMA_UNROLL
-    for (short i = 0; i < kElemsPerFrag; i++) {
-      ct_a[i] = A[i];
-    }
-
-    // Load B into right operand registers
-    STEEL_PRAGMA_UNROLL
-    for (short i = 0; i < kElemsPerFrag; i++) {
-      ct_b[i] = Bn0[i];
-      ct_b[kElemsPerFrag + i] = Bn1[i];
-    }
-
-    // Load C into output registers (op handles accumulation)
-    STEEL_PRAGMA_UNROLL
-    for (short i = 0; i < kElemsPerFrag; i++) {
-      ct_c[i] = Cn0[i];
-      ct_c[kElemsPerFrag + i] = Cn1[i];
-    }
-
-    // Do matmul
-    gemm_op.run(ct_a, ct_b, ct_c);
-
-    // Copy out results
-    STEEL_PRAGMA_UNROLL
-    for (short i = 0; i < kElemsPerFrag; i++) {
-      Cn0[i] = ct_c[i];
-      Cn1[i] = ct_c[kElemsPerFrag + i];
-    }
-  }
-
-  template <
-      typename CType,
-      typename AType,
-      typename BType,
-      bool transpose_a = false,
-      bool transpose_b = false>
-  METAL_FUNC static constexpr void mma(
-      thread dtype_frag_t<CType>& Cm0,
-      thread dtype_frag_t<CType>& Cm1,
-      const thread dtype_frag_t<AType>& Am0,
-      const thread dtype_frag_t<AType>& Am1,
-      metal::bool_constant<transpose_a>,
-      const thread dtype_frag_t<BType>& B,
-      metal::bool_constant<transpose_b>) {
-    // Create Matmul descriptor
-    constexpr auto desc = mpp::tensor_ops::matmul2d_descriptor(
-        16,
-        32,
-        16,
-        transpose_a,
-        transpose_b,
-        true,
-        mpp::tensor_ops::matmul2d_descriptor::mode::multiply_accumulate);
-
-    // Create matmul op
-    mpp::tensor_ops::matmul2d<desc, metal::execution_simdgroup> gemm_op;
-
-    // Create matmul operands in registers
-    auto ct_a =
-        gemm_op
-            .template get_left_input_cooperative_tensor<AType, BType, CType>();
-    auto ct_b =
-        gemm_op
-            .template get_right_input_cooperative_tensor<AType, BType, CType>();
-
-    // Create matmul output in register
-    auto ct_c = gemm_op.template get_destination_cooperative_tensor<
-        decltype(ct_a),
-        decltype(ct_b),
-        CType>();
-
-    // Load A in to left operand registers
-    STEEL_PRAGMA_UNROLL
-    for (short i = 0; i < kElemsPerFrag; i++) {
-      ct_a[i] = Am0[i];
-      ct_a[kElemsPerFrag + i] = Am1[i];
-    }
-
-    // Load B into right operand registers
-    STEEL_PRAGMA_UNROLL
-    for (short i = 0; i < kElemsPerFrag; i++) {
-      ct_b[i] = B[i];
-    }
-
-    // Load C into output registers (op handles accumulation)
-    STEEL_PRAGMA_UNROLL
-    for (short i = 0; i < kElemsPerFrag; i++) {
-      ct_c[i] = Cm0[i];
-      ct_c[kElemsPerFrag + i] = Cm1[i];
-    }
-
-    // Do matmul
-    gemm_op.run(ct_a, ct_b, ct_c);
-
-    // Copy out results
-    STEEL_PRAGMA_UNROLL
-    for (short i = 0; i < kElemsPerFrag; i++) {
-      Cm0[i] = ct_c[i];
-      Cm1[i] = ct_c[kElemsPerFrag + i];
-    }
-  }
-};
-
-template <
-    typename T,
-    short kTileRows_,
-    short kTileCols_,
-    class NAXFrag_ = BaseNAXFrag>
-struct NAXTile {
-  using NAXFrag_t = NAXFrag_;
-  using elem_type = T;
-
-  STEEL_CONST short kFragRows = NAXFrag_t::kFragRows;
-  STEEL_CONST short kFragCols = NAXFrag_t::kFragCols;
-  STEEL_CONST short kElemsPerFrag = NAXFrag_t::kElemsPerFrag;
-
-  STEEL_CONST short kTileRows = kTileRows_;
-  STEEL_CONST short kTileCols = kTileCols_;
-
-  STEEL_CONST short kRows = kTileRows * kFragRows;
-  STEEL_CONST short kCols = kTileCols * kFragCols;
-
-  STEEL_CONST short kNumFrags = kTileRows * kTileCols;
-  STEEL_CONST short kElemsPerTile = kNumFrags * kElemsPerFrag;
-
-  STEEL_CONST short kFragThrRows = NAXFrag_t::kElemRows;
-  STEEL_CONST short kFragThrCols = NAXFrag_t::kElemCols;
-  STEEL_CONST short kFragRowsJump = NAXFrag_t::kElemRowsJump;
-
-  STEEL_CONST short kRowsPerThread = kTileRows * NAXFrag_t::kElemRows;
-  STEEL_CONST short kColsPerThread = kTileCols * NAXFrag_t::kElemCols;
-
-  typedef typename NAXFrag_t::template dtype_frag_t<T> frag_type;
-
-  frag_type val_frags[kNumFrags]; // = {frag_type(0)};
-
-  METAL_FUNC NAXTile() thread {}
-
-  METAL_FUNC constexpr void clear() {
-    STEEL_PRAGMA_UNROLL
-    for (short i = 0; i < kNumFrags; ++i) {
-      val_frags[i] = frag_type(0);
-    }
-  }
-
-  METAL_FUNC constexpr thread frag_type& frag_at(const short i, const short j) {
-    return val_frags[i * kTileCols + j];
-  }
-
-  METAL_FUNC constexpr const thread frag_type& frag_at(
-      const short i,
-      const short j) const {
-    return val_frags[i * kTileCols + j];
-  }
-
-  template <int i, int j>
-  METAL_FUNC constexpr thread frag_type& frag_at() {
-    return val_frags[i * kTileCols + j];
-  }
-
-  template <int i, int j>
-  METAL_FUNC constexpr const thread frag_type& frag_at() const {
-    return val_frags[i * kTileCols + j];
-  }
-
-  template <bool transpose>
-  METAL_FUNC constexpr thread frag_type&
-  frag_at(const short i, const short j, metal::bool_constant<transpose>) {
-    if constexpr (transpose) {
-      return frag_at(j, i);
-    } else {
-      return frag_at(i, j);
-    }
-  }
-
-  template <bool transpose>
-  METAL_FUNC constexpr const thread frag_type&
-  frag_at(const short i, const short j, metal::bool_constant<transpose>) const {
-    if constexpr (transpose) {
-      return frag_at(j, i);
-    } else {
-      return frag_at(i, j);
-    }
-  }
-
-  template <int i, int j, bool transpose>
-  METAL_FUNC constexpr thread frag_type& frag_at() {
-    if constexpr (transpose) {
-      return frag_at<j, i>();
-    } else {
-      return frag_at<i, j>();
-    }
-  }
-
-  template <int i, int j, bool transpose>
-  METAL_FUNC constexpr const thread frag_type& frag_at() const {
-    if constexpr (transpose) {
-      return frag_at<j, i>();
-    } else {
-      return frag_at<i, j>();
-    }
-  }
-
-  METAL_FUNC thread elem_type* elems() {
-    return reinterpret_cast<thread elem_type*>(val_frags);
-  }
-
-  METAL_FUNC const thread elem_type* elems() const {
-    return reinterpret_cast<const thread elem_type*>(val_frags);
-  }
-
-  template <typename Op>
-  METAL_FUNC void row_reduce(thread metal::vec<T, kRowsPerThread>& vals) const {
-    auto vptr = (thread T*)(&vals);
-    STEEL_PRAGMA_UNROLL
-    for (short i = 0; i < kTileRows; ++i) {
-      STEEL_PRAGMA_UNROLL
-      for (short j = 0; j < kTileCols; ++j) {
-        NAXFrag_t::template row_reduce<Op>(
-            frag_at(i, j), &vptr[i * kFragThrRows]);
-      }
-    }
-  }
-
-  template <typename Op>
-  METAL_FUNC void row_bin_op(thread metal::vec<T, kRowsPerThread>& vals) {
-    auto vptr = (thread T*)(&vals);
-    STEEL_PRAGMA_UNROLL
-    for (short i = 0; i < kTileRows; ++i) {
-      STEEL_PRAGMA_UNROLL
-      for (short j = 0; j < kTileCols; ++j) {
-        NAXFrag_t::template row_bin_op<Op>(
-            frag_at(i, j), &vptr[i * kFragThrRows]);
-      }
-    }
-  }
-
-  template <typename U, int str_x, int str_y>
-  METAL_FUNC void load(const threadgroup U* src) {
-    const_for_loop<0, kTileRows, 1>([&](auto idx_row) {
-      const_for_loop<0, kTileCols, 1>([&](auto idx_col) {
-        NAXFrag_t::load(
-            frag_at<idx_row.value, idx_col.value>(),
-            src,
-            Int<str_x>{},
-            Int<str_y>{},
-            idx_row * Int<kFragRows>{},
-            idx_col * Int<kFragCols>{});
-      });
-    });
-  }
-
-  template <typename U, int str_x, int str_y>
-  METAL_FUNC void store(threadgroup U* dst) const {
-    const_for_loop<0, kTileRows, 1>([&](auto idx_row) {
-      const_for_loop<0, kTileCols, 1>([&](auto idx_col) {
-        NAXFrag_t::store(
-            frag_at<idx_row.value, idx_col.value>(),
-            dst,
-            Int<str_x>{},
-            Int<str_y>{},
-            idx_row * Int<kFragRows>{},
-            idx_col * Int<kFragCols>{});
-      });
-    });
-  }
-
-  template <typename U>
-  METAL_FUNC void load(const device U* src, const int ld) {
-    const_for_loop<0, kTileRows, 1>([&](auto idx_row) {
-      const_for_loop<0, kTileCols, 1>([&](auto idx_col) {
-        NAXFrag_t::load(
-            frag_at<idx_row.value, idx_col.value>(),
-            src,
-            ld,
-            Int<1>{},
-            idx_row * Int<kFragRows>{},
-            idx_col * Int<kFragCols>{});
-      });
-    });
-  }
-
-  template <typename U>
-  METAL_FUNC void store(device U* dst, const int ld) const {
-    const_for_loop<0, kTileRows, 1>([&](auto idx_row) {
-      const_for_loop<0, kTileCols, 1>([&](auto idx_col) {
-        NAXFrag_t::store(
-            frag_at<idx_row.value, idx_col.value>(),
-            dst,
-            ld,
-            Int<1>{},
-            idx_row * Int<kFragRows>{},
-            idx_col * Int<kFragCols>{});
-      });
-    });
-  }
-
-  template <typename U>
-  METAL_FUNC void
-  load_rows(const device U* src, const int ld, const short n_rows) {
-    const_for_loop<0, kTileRows, 1>([&](auto idx_row) {
-      const_for_loop<0, kTileCols, 1>([&](auto idx_col) {
-        NAXFrag_t::load_rows(
-            frag_at<idx_row.value, idx_col.value>(),
-            src,
-            ld,
-            Int<1>{},
-            n_rows,
-            idx_row * Int<kFragRows>{},
-            idx_col * Int<kFragCols>{});
-      });
-    });
-  }
-
-  template <typename U>
-  METAL_FUNC void
-  load_safe(const device U* src, const int ld, const short2 src_tile_dims) {
-    const_for_loop<0, kTileRows, 1>([&](auto idx_row) {
-      const_for_loop<0, kTileCols, 1>([&](auto idx_col) {
-        NAXFrag_t::load_safe(
-            frag_at<idx_row.value, idx_col.value>(),
-            src,
-            ld,
-            Int<1>{},
-            src_tile_dims.y,
-            src_tile_dims.x,
-            idx_row * Int<kFragRows>{},
-            idx_col * Int<kFragCols>{});
-      });
-    });
-  }
-
-  template <typename U>
-  METAL_FUNC void store_rows(device U* dst, const int ld, const short n_rows)
-      const {
-    const_for_loop<0, kTileRows, 1>([&](auto idx_row) {
-      const_for_loop<0, kTileCols, 1>([&](auto idx_col) {
-        NAXFrag_t::store_rows(
-            frag_at<idx_row.value, idx_col.value>(),
-            dst,
-            ld,
-            Int<1>{},
-            n_rows,
-            idx_row * Int<kFragRows>{},
-            idx_col * Int<kFragCols>{});
-      });
-    });
-  }
-
-  template <typename U>
-  METAL_FUNC void
-  store_safe(device U* dst, const int ld, const short2 dst_tile_dims) const {
-    const_for_loop<0, kTileRows, 1>([&](auto idx_row) {
-      const_for_loop<0, kTileCols, 1>([&](auto idx_col) {
-        NAXFrag_t::store_safe(
-            frag_at<idx_row.value, idx_col.value>(),
-            dst,
-            ld,
-            Int<1>{},
-            dst_tile_dims.y,
-            dst_tile_dims.x,
-            idx_row * Int<kFragRows>{},
-            idx_col * Int<kFragCols>{});
-      });
-    });
-  }
-
-  template <typename U>
-  METAL_FUNC void store_slice(
-      device U* dst,
-      const int ld,
-      const short2 start,
-      const short2 stop) const {
-    const_for_loop<0, kTileRows, 1>([&](auto idx_row) {
-      const_for_loop<0, kTileCols, 1>([&](auto idx_col) {
-        NAXFrag_t::store_slice(
-            frag_at<idx_row.value, idx_col.value>(),
-            dst,
-            ld,
-            Int<1>{},
-            start.y,
-            stop.y,
-            start.x,
-            stop.x,
-            idx_row * Int<kFragRows>{},
-            idx_col * Int<kFragCols>{});
-      });
-    });
-  }
-};
-
-template <
-    class CTile,
-    class ATile,
-    class BTile,
-    bool transpose_a,
-    bool transpose_b>
-METAL_FUNC void tile_matmad_nax(
-    thread CTile& C,
-    thread ATile& A,
-    metal::bool_constant<transpose_a>,
-    thread BTile& B,
-    metal::bool_constant<transpose_b>) {
-  // Static checks
-  constexpr short TMa = transpose_a ? ATile::kTileCols : ATile::kTileRows;
-  constexpr short TM = CTile::kTileRows;
-  static_assert(TMa == TM, "MXU tile matmul: M dimensions do not match");
-
-  constexpr short TNb = transpose_b ? BTile::kTileRows : BTile::kTileCols;
-  constexpr short TN = CTile::kTileCols;
-  static_assert(TNb == TN, "MXU tile matmul: N dimensions do not match");
-
-  constexpr short TKa = transpose_a ? ATile::kTileRows : ATile::kTileCols;
-  constexpr short TK = transpose_b ? BTile::kTileCols : BTile::kTileRows;
-  static_assert(TKa == TK, "MXU tile matmul: K dimensions do not match");
-
-  constexpr auto ta = metal::bool_constant<transpose_a>{};
-  constexpr auto tb = metal::bool_constant<transpose_b>{};
-
-  if constexpr (TN == 1 && TM % 2 == 0) {
-    STEEL_PRAGMA_UNROLL
-    for (short mm = 0; mm < TM; mm += 2) {
-      STEEL_PRAGMA_UNROLL
-      for (short nn = 0; nn < TN; ++nn) {
-        STEEL_PRAGMA_UNROLL
-        for (short kk = 0; kk < TK; ++kk) {
-          CTile::NAXFrag_t::mma(
-              C.frag_at(mm, nn),
-              C.frag_at(mm + 1, nn),
-              A.frag_at(mm, kk, ta),
-              A.frag_at(mm + 1, kk, ta),
-              metal::bool_constant<transpose_a>{},
-              B.frag_at(kk, nn, tb),
-              metal::bool_constant<transpose_b>{});
-        }
-      }
-    }
-  } else if constexpr (TN % 2 == 0) {
-    STEEL_PRAGMA_UNROLL
-    for (short mm = 0; mm < TM; ++mm) {
-      STEEL_PRAGMA_UNROLL
-      for (short nn = 0; nn < TN; nn += 2) {
-        STEEL_PRAGMA_UNROLL
-        for (short kk = 0; kk < TK; ++kk) {
-          CTile::NAXFrag_t::mma(
-              C.frag_at(mm, nn),
-              C.frag_at(mm, nn + 1),
-              A.frag_at(mm, kk, ta),
-              metal::bool_constant<transpose_a>{},
-              B.frag_at(kk, nn, tb),
-              B.frag_at(kk, nn + 1, tb),
-              metal::bool_constant<transpose_b>{});
-        }
-      }
-    }
-  }
-}
-
-} // namespace steel
-} // namespace mlx
+// attn-specific NAX extensions go here. None today.
