+++ mlx/backend/metal/kernels/steel/nax_common.h
@@ -1,0 +1,1605 @@
+// Copyright © 2024-26 Apple Inc.
+//
+// Shared NAX class hierarchy used by both the gemm and attn trees.
+// This is the canonical definition of Role, BaseNAXFrag, NAXFrag32,
+// NAXTile, and tile_matmad_nax.  Tree-specific headers (gemm/nax.h and
+// attn/nax.h) re-export from here; all new NAX code should go here.
+
+#pragma once
+
+#include <metal_simdgroup>
+#include <metal_simdgroup_matrix>
+#include <metal_stdlib>
+
+#include "mlx/backend/metal/kernels/steel/defines.h"
+#include "mlx/backend/metal/kernels/steel/utils/integral_constant.h"
+
+#include <MetalPerformancePrimitives/MetalPerformancePrimitives.h>
+
+using namespace metal;
+
+///////////////////////////////////////////////////////////////////////////////
+// MMA helper
+///////////////////////////////////////////////////////////////////////////////
+
+namespace mlx {
+namespace steel {
+
+// Role of an operand in a matmul2d descriptor.
+// Left  → A input (transpose_left  flag in the descriptor).
+// Right → B input (transpose_right flag in the descriptor).
+enum class Role { Left, Right };
+
+///////////////////////////////////////////////////////////////////////////////
+// NAX Steel with new tiles
+///////////////////////////////////////////////////////////////////////////////
+
+struct BaseNAXFrag {
+  STEEL_CONST short kFragRows = 16;
+  STEEL_CONST short kFragCols = 16;
+
+  STEEL_CONST short kElemsPerFrag = (kFragRows * kFragCols) / 32;
+
+  STEEL_CONST short kElemRows = 2;
+  STEEL_CONST short kElemCols = 4;
+
+  STEEL_CONST short kElemRowsJump = 8;
+
+  // Number of frags produced per MMA along N. The (16, 32, 16) descriptor
+  // produces a 16x32 output = 2 frags wide. tile_matmad_nax dispatches on this.
+  STEEL_CONST short kPacking = 2;
+
+  static_assert(
+      kElemRows * kElemCols == kElemsPerFrag,
+      "MMAFrag shape is not consistent with MMAFrag size");
+
+  template <typename U>
+  using dtype_frag_t = typename metal::vec<U, kElemsPerFrag>;
+
+  METAL_FUNC static short2 get_coord() {
+    const ushort simd_lane_id = __metal_get_thread_index_in_simdgroup(ushort());
+    const short qid = simd_lane_id >> 2;
+    const short fm = ((qid & 4) | ((simd_lane_id >> 1) & 3));
+    const short fn = ((qid & 2) | (simd_lane_id & 1)) * 4;
+    return short2{fn, fm};
+  }
+
+  METAL_FUNC static short2 get_coord(short idx) {
+    const ushort simd_lane_id = __metal_get_thread_index_in_simdgroup(ushort());
+    const short qid = simd_lane_id >> 2;
+    const short fm = ((qid & 4) | ((simd_lane_id >> 1) & 3)) + (idx >> 2) * 8;
+    const short fn = ((qid & 2) | (simd_lane_id & 1)) * 4 + idx % 4;
+    return short2{fn, fm};
+  }
+
+  template <
+      typename T,
+      typename SrcPtrType,
+      typename StrX,
+      typename StrY,
+      typename OffX = Int<0>,
+      typename OffY = Int<0>>
+  METAL_FUNC static constexpr void load(
+      thread dtype_frag_t<T>& dst,
+      SrcPtrType src,
+      StrX str_x,
+      StrY str_y,
+      OffX off_x = {},
+      OffY off_y = {}) {
+    const short2 sc = get_coord();
+    src += sc.y * str_x + sc.x * str_y;
+
+    STEEL_PRAGMA_UNROLL
+    for (short i = 0; i < kElemRows; i++) {
+      const auto r = off_x + i * kElemRowsJump;
+      const auto c = off_y;
+
+      if constexpr (metal::is_same_v<StrY, Int<1>>) {
+        STEEL_PRAGMA_UNROLL
+        for (short j = 0; j < kElemCols; j++) {
+          dst[i * kElemCols + j] = static_cast<T>(src[r * str_x + c + j]);
+        }
+      } else {
+        STEEL_PRAGMA_UNROLL
+        for (short j = 0; j < kElemCols; j++) {
+          dst[i * kElemCols + j] =
+              static_cast<T>(src[r * str_x + (c + j) * str_y]);
+        }
+      }
+    }
+  }
+
+  template <
+      typename T,
+      typename SrcPtrType,
+      typename StrX,
+      typename StrY,
+      typename LimX,
+      typename OffX = Int<0>,
+      typename OffY = Int<0>>
+  METAL_FUNC static constexpr void load_rows(
+      thread dtype_frag_t<T>& dst,
+      SrcPtrType src,
+      StrX str_x,
+      StrY str_y,
+      LimX lim_x,
+      OffX off_x = {},
+      OffY off_y = {}) {
+    const short2 sc = get_coord();
+    src += sc.y * str_x + sc.x * str_y;
+    auto lx = lim_x - sc.y;
+
+    STEEL_PRAGMA_UNROLL
+    for (short i = 0; i < kElemRows; i++) {
+      const auto r = off_x + i * kElemRowsJump;
+      const auto c = off_y;
+
+      if (r < lx) {
+        if constexpr (metal::is_same_v<StrY, Int<1>>) {
+          STEEL_PRAGMA_UNROLL
+          for (short j = 0; j < kElemCols; j++) {
+            dst[i * kElemCols + j] = static_cast<T>(src[r * str_x + (c + j)]);
+          }
+        } else {
+          STEEL_PRAGMA_UNROLL
+          for (short j = 0; j < kElemCols; j++) {
+            dst[i * kElemCols + j] =
+                static_cast<T>(src[r * str_x + (c + j) * str_y]);
+          }
+        }
+
+      } else {
+        STEEL_PRAGMA_UNROLL
+        for (short j = 0; j < kElemCols; j++) {
+          dst[i * kElemCols + j] = T(0);
+        }
+      }
+    }
+  }
+
+  template <
+      typename T,
+      typename SrcPtrType,
+      typename StrX,
+      typename StrY,
+      typename LimX,
+      typename LimY,
+      typename OffX = Int<0>,
+      typename OffY = Int<0>>
+  METAL_FUNC static constexpr void load_safe(
+      thread dtype_frag_t<T>& dst,
+      SrcPtrType src,
+      StrX str_x,
+      StrY str_y,
+      LimX lim_x,
+      LimY lim_y,
+      OffX off_x = {},
+      OffY off_y = {}) {
+    const short2 sc = get_coord();
+    src += sc.y * str_x + sc.x * str_y;
+    auto lx = lim_x - sc.y;
+    auto ly = lim_y - sc.x;
+
+    STEEL_PRAGMA_UNROLL
+    for (short i = 0; i < kElemRows; i++) {
+      const auto r = off_x + i * kElemRowsJump;
+      const auto c = off_y;
+      STEEL_PRAGMA_UNROLL
+      for (short j = 0; j < kElemCols; j++) {
+        if ((r < lx) && ((c + j) < ly)) {
+          dst[i * kElemCols + j] =
+              static_cast<T>(src[r * str_x + (c + j) * str_y]);
+        } else {
+          dst[i * kElemCols + j] = T(0);
+        }
+      }
+    }
+  }
+
+  template <
+      typename T,
+      typename DstPtrType,
+      typename StrX,
+      typename StrY,
+      typename OffX = Int<0>,
+      typename OffY = Int<0>>
+  METAL_FUNC static constexpr void store(
+      const thread dtype_frag_t<T>& src,
+      DstPtrType dst,
+      StrX str_x,
+      StrY str_y,
+      OffX off_x = {},
+      OffY off_y = {}) {
+    using U = pointer_element_t<DstPtrType>;
+
+    const short2 sc = get_coord();
+    dst += sc.y * str_x + sc.x * str_y;
+
+    STEEL_PRAGMA_UNROLL
+    for (short i = 0; i < kElemRows; i++) {
+      const auto r = off_x + i * kElemRowsJump;
+      const auto c = off_y;
+
+      if constexpr (metal::is_same_v<StrY, Int<1>>) {
+        STEEL_PRAGMA_UNROLL
+        for (short j = 0; j < kElemCols; j++) {
+          dst[r * str_x + c + j] = static_cast<U>(src[i * kElemCols + j]);
+        }
+      } else {
+        STEEL_PRAGMA_UNROLL
+        for (short j = 0; j < kElemCols; j++) {
+          dst[r * str_x + (c + j) * str_y] =
+              static_cast<U>(src[i * kElemCols + j]);
+        }
+      }
+    }
+  }
+
+  template <
+      typename T,
+      typename DstPtrType,
+      typename StrX,
+      typename StrY,
+      typename LimX,
+      typename OffX = Int<0>,
+      typename OffY = Int<0>>
+  METAL_FUNC static constexpr void store_rows(
+      const thread dtype_frag_t<T>& src,
+      DstPtrType dst,
+      StrX str_x,
+      StrY str_y,
+      LimX lim_x,
+      OffX off_x = {},
+      OffY off_y = {}) {
+    using U = pointer_element_t<DstPtrType>;
+
+    const short2 sc = get_coord();
+    dst += sc.y * str_x + sc.x * str_y;
+    auto lx = lim_x - sc.y;
+
+    STEEL_PRAGMA_UNROLL
+    for (short i = 0; i < kElemRows; i++) {
+      const auto r = off_x + i * kElemRowsJump;
+      const auto c = off_y;
+
+      if (r < lx) {
+        if constexpr (metal::is_same_v<StrY, Int<1>>) {
+          STEEL_PRAGMA_UNROLL
+          for (short j = 0; j < kElemCols; j++) {
+            dst[r * str_x + c + j] = static_cast<U>(src[i * kElemCols + j]);
+          }
+        } else {
+          STEEL_PRAGMA_UNROLL
+          for (short j = 0; j < kElemCols; j++) {
+            dst[r * str_x + (c + j) * str_y] =
+                static_cast<U>(src[i * kElemCols + j]);
+          }
+        }
+      }
+    }
+  }
+
+  template <
+      typename T,
+      typename DstPtrType,
+      typename StrX,
+      typename StrY,
+      typename LimX,
+      typename LimY,
+      typename OffX = Int<0>,
+      typename OffY = Int<0>>
+  METAL_FUNC static constexpr void store_safe(
+      const thread dtype_frag_t<T>& src,
+      DstPtrType dst,
+      StrX str_x,
+      StrY str_y,
+      LimX lim_x,
+      LimY lim_y,
+      OffX off_x = {},
+      OffY off_y = {}) {
+    using U = pointer_element_t<DstPtrType>;
+
+    const short2 sc = get_coord();
+    dst += sc.y * str_x + sc.x * str_y;
+    auto lx = lim_x - sc.y;
+    auto ly = lim_y - sc.x;
+
+    STEEL_PRAGMA_UNROLL
+    for (short i = 0; i < kElemRows; i++) {
+      const auto r = off_x + i * kElemRowsJump;
+      const auto c = off_y;
+
+      STEEL_PRAGMA_UNROLL
+      for (short j = 0; j < kElemCols; j++) {
+        if (r < lx && (c + j) < ly) {
+          dst[r * str_x + (c + j) * str_y] =
+              static_cast<U>(src[i * kElemCols + j]);
+        }
+      }
+    }
+  }
+
+  template <
+      typename T,
+      typename DstPtrType,
+      typename StrX,
+      typename StrY,
+      typename StartX,
+      typename StopX,
+      typename StartY,
+      typename StopY,
+      typename OffX = Int<0>,
+      typename OffY = Int<0>>
+  METAL_FUNC static constexpr void store_slice(
+      const thread dtype_frag_t<T>& src,
+      DstPtrType dst,
+      StrX str_x,
+      StrY str_y,
+      StartX start_x,
+      StopX stop_x,
+      StartY start_y,
+      StopY stop_y,
+      OffX off_x = Int<0>{},
+      OffY off_y = Int<0>{}) {
+    using U = pointer_element_t<DstPtrType>;
+
+    const short2 sc = get_coord();
+
+    const_for_loop<0, kElemRows, 1>([&](auto idx_row) {
+      const auto r = off_x + idx_row * Int<kElemRowsJump>{};
+      if (r >= stop_x - sc.y || r < start_x - sc.y) {
+        return;
+      }
+
+      const_for_loop<0, kElemCols, 1>([&](auto idx_col) {
+        const auto c = off_y + idx_col;
+        if (c >= stop_y - sc.x || c < start_y - sc.x) {
+          return;
+        }
+
+        const auto src_idx = idx_row * Int<kElemCols>{} + idx_col;
+        dst[(r + sc.y) * str_x + (c + sc.x) * str_y] =
+            static_cast<U>(src[src_idx]);
+      });
+    });
+  }
+
+  template <typename Op, typename T>
+  METAL_FUNC static constexpr void row_reduce(
+      thread const dtype_frag_t<T>& inp_vals,
+      thread T* reduced_vals) {
+    STEEL_PRAGMA_UNROLL
+    for (short i = 0; i < kElemRows; i++) {
+      T thr_reduce = Op::apply(
+          Op::apply(inp_vals[i * kElemCols + 0], inp_vals[i * kElemCols + 1]),
+          Op::apply(inp_vals[i * kElemCols + 2], inp_vals[i * kElemCols + 3]));
+
+      T qgr_reduce = simd_shuffle_xor(thr_reduce, ushort(1));
+      qgr_reduce = Op::apply(thr_reduce, qgr_reduce);
+
+      T sgr_reduce = simd_shuffle_xor(qgr_reduce, ushort(8));
+      sgr_reduce = Op::apply(qgr_reduce, sgr_reduce);
+
+      reduced_vals[i] = Op::apply(reduced_vals[i], sgr_reduce);
+    }
+  }
+
+  template <typename Op, typename T>
+  METAL_FUNC static constexpr void row_bin_op(
+      thread dtype_frag_t<T>& inp_vals,
+      thread T* row_vals) {
+    STEEL_PRAGMA_UNROLL
+    for (short i = 0; i < kElemRows; i++) {
+      STEEL_PRAGMA_UNROLL
+      for (short j = 0; j < kElemCols; j++) {
+        inp_vals[i * kElemCols + j] =
+            Op::apply(inp_vals[i * kElemCols + j], row_vals[i]);
+      }
+    }
+  }
+
+  template <
+      typename CType,
+      typename AType,
+      typename BType,
+      bool transpose_a = false,
+      bool transpose_b = false>
+  METAL_FUNC static constexpr void mma(
+      thread dtype_frag_t<CType>& Cn0,
+      thread dtype_frag_t<CType>& Cn1,
+      const thread dtype_frag_t<AType>& A,
+      metal::bool_constant<transpose_a>,
+      const thread dtype_frag_t<BType>& Bn0,
+      const thread dtype_frag_t<BType>& Bn1,
+      metal::bool_constant<transpose_b>) {
+    constexpr auto desc = mpp::tensor_ops::matmul2d_descriptor(
+        16,
+        32,
+        16,
+        transpose_a,
+        transpose_b,
+        true,
+        mpp::tensor_ops::matmul2d_descriptor::mode::multiply_accumulate);
+
+    // Create matmul op
+    mpp::tensor_ops::matmul2d<desc, metal::execution_simdgroup> gemm_op;
+
+    // Create matmul operands in registers
+    auto ct_a =
+        gemm_op
+            .template get_left_input_cooperative_tensor<AType, BType, CType>();
+    auto ct_b =
+        gemm_op
+            .template get_right_input_cooperative_tensor<AType, BType, CType>();
+
+    // Create matmul output in register
+    auto ct_c = gemm_op.template get_destination_cooperative_tensor<
+        decltype(ct_a),
+        decltype(ct_b),
+        CType>();
+
+    // Load A in to left operand registers
+    STEEL_PRAGMA_UNROLL
+    for (short i = 0; i < kElemsPerFrag; i++) {
+      ct_a[i] = A[i];
+    }
+
+    // Load B into right operand registers
+    STEEL_PRAGMA_UNROLL
+    for (short i = 0; i < kElemsPerFrag; i++) {
+      ct_b[i] = Bn0[i];
+      ct_b[kElemsPerFrag + i] = Bn1[i];
+    }
+
+    // Load C into output registers (op handles accumulation)
+    STEEL_PRAGMA_UNROLL
+    for (short i = 0; i < kElemsPerFrag; i++) {
+      ct_c[i] = Cn0[i];
+      ct_c[kElemsPerFrag + i] = Cn1[i];
+    }
+
+    // Do matmul
+    gemm_op.run(ct_a, ct_b, ct_c);
+
+    // Copy out results
+    STEEL_PRAGMA_UNROLL
+    for (short i = 0; i < kElemsPerFrag; i++) {
+      Cn0[i] = ct_c[i];
+      Cn1[i] = ct_c[kElemsPerFrag + i];
+    }
+  }
+
+  template <
+      typename CType,
+      typename AType,
+      typename BType,
+      bool transpose_a = false,
+      bool transpose_b = false>
+  METAL_FUNC static constexpr void mma(
+      thread dtype_frag_t<CType>& Cm0,
+      thread dtype_frag_t<CType>& Cm1,
+      const thread dtype_frag_t<AType>& Am0,
+      const thread dtype_frag_t<AType>& Am1,
+      metal::bool_constant<transpose_a>,
+      const thread dtype_frag_t<BType>& B,
+      metal::bool_constant<transpose_b>) {
+    // Create Matmul descriptor
+    constexpr auto desc = mpp::tensor_ops::matmul2d_descriptor(
+        16,
+        32,
+        16,
+        transpose_a,
+        transpose_b,
+        true,
+        mpp::tensor_ops::matmul2d_descriptor::mode::multiply_accumulate);
+
+    // Create matmul op
+    mpp::tensor_ops::matmul2d<desc, metal::execution_simdgroup> gemm_op;
+
+    // Create matmul operands in registers
+    auto ct_a =
+        gemm_op
+            .template get_left_input_cooperative_tensor<AType, BType, CType>();
+    auto ct_b =
+        gemm_op
+            .template get_right_input_cooperative_tensor<AType, BType, CType>();
+
+    // Create matmul output in register
+    auto ct_c = gemm_op.template get_destination_cooperative_tensor<
+        decltype(ct_a),
+        decltype(ct_b),
+        CType>();
+
+    // Load A in to left operand registers
+    STEEL_PRAGMA_UNROLL
+    for (short i = 0; i < kElemsPerFrag; i++) {
+      ct_a[i] = Am0[i];
+      ct_a[kElemsPerFrag + i] = Am1[i];
+    }
+
+    // Load B into right operand registers
+    STEEL_PRAGMA_UNROLL
+    for (short i = 0; i < kElemsPerFrag; i++) {
+      ct_b[i] = B[i];
+    }
+
+    // Load C into output registers (op handles accumulation)
+    STEEL_PRAGMA_UNROLL
+    for (short i = 0; i < kElemsPerFrag; i++) {
+      ct_c[i] = Cm0[i];
+      ct_c[kElemsPerFrag + i] = Cm1[i];
+    }
+
+    // Do matmul
+    gemm_op.run(ct_a, ct_b, ct_c);
+
+    // Copy out results
+    STEEL_PRAGMA_UNROLL
+    for (short i = 0; i < kElemsPerFrag; i++) {
+      Cm0[i] = ct_c[i];
+      Cm1[i] = ct_c[kElemsPerFrag + i];
+    }
+  }
+};
+
+///////////////////////////////////////////////////////////////////////////////
+// NAXFrag32 — 32x32 frag, single-frag MMA, layout-agnostic I/O via
+// metal::tensor_inline. Used on g16 (M4 Pro), where matmul2d is only correct
+// for the (32, 32, 32) descriptor.
+//
+// I/O strategy (chosen during Task 1.5 of the NAX g16 fix):
+//   The Metal SDK exposes only `tensor_inline`, `tensor_handle`, and
+//   `tensor_offset` access kinds — no `tensor_padded` or strided/bounded
+//   variants. Cooperative tensors have no masked-load entry points. So
+//   "Strategy A" (route safe/rows variants through an MPP-bounded view) is
+//   infeasible on this SDK. We use "Strategy B": the safe/rows variants
+//   stage through a 32x32 threadgroup scratch buffer with a cooperative
+//   bounds-checked copy, then route through the contiguous load/store.
+//
+//   Metal forbids declaring `threadgroup T buf[N]` inside non-kernel
+//   device functions. So load_safe/load_rows/store_safe/store_rows take an
+//   extra `threadgroup T* scratch` parameter — callers must allocate
+//   `threadgroup T scratch[32 * 32]` at kernel scope and pass it in.
+//
+//   This is a real signature divergence from BaseNAXFrag. NAXTile's
+//   wrappers (load_rows, load_safe, store_rows, store_safe, store_slice)
+//   currently follow BaseNAXFrag's convention; Phase 2 / Task 2.2 must
+//   redesign that dispatch before any NAXTile<..., NAXFrag32> instantiation
+//   compiles. See docs/superpowers/specs/2026-04-27-nax-g16-fix-plan.md.
+///////////////////////////////////////////////////////////////////////////////
+
+struct NAXFrag32 {
+  STEEL_CONST short kFragRows = 32;
+  STEEL_CONST short kFragCols = 32;
+
+  STEEL_CONST short kElemsPerFrag = (kFragRows * kFragCols) / 32;  // 32
+
+  STEEL_CONST short kElemRows = 4;
+  STEEL_CONST short kElemCols = 8;
+
+  // NAXFrag32's per-thread row layout is non-uniform (dr_table =
+  // {0,1,8,9,16,17,24,25}) — there is no valid constant stride. We set this
+  // to 1 only because NAXTile propagates the field as kFragRowsJump; do NOT
+  // use it for stride arithmetic. Callers that need per-element offsets
+  // must call dr_dc() explicitly. See Phase 5 (SDPA) of
+  // docs/superpowers/specs/2026-04-27-nax-g16-fix-plan.md for the
+  // attention path that currently uses this stride and will need rework.
+  STEEL_CONST short kElemRowsJump = 1;
+
+  // One MMA produces one full 32x32 frag, i.e. unpacked.
+  STEEL_CONST short kPacking = 1;
+
+  static_assert(
+      kElemRows * kElemCols == kElemsPerFrag,
+      "NAXFrag32 shape is not consistent with its size");
+
+  template <typename U>
+  using dtype_frag_t = typename metal::vec<U, kElemsPerFrag>;
+
+  // Per-thread base coordinate in the 32x32 frag. Element i lives at
+  //   (b + dc[i % 8], a + dr[i / 8])  // {col, row} — matches short2 / dr_dc()
+  // with
+  //   a (row base, 8 values in {0..7}) = ((lane & 2) >> 1) | ((lane & 4) >> 1) | ((lane & 16) >> 2)
+  //   b (col base, 4 values in {0,2,4,6}) = ((lane & 1) << 1) | ((lane & 8) >> 1)
+  //   dr = {0, 8, 16, 24}                          // 4 row offsets
+  //   dc = {0, 1, 8, 9, 16, 17, 24, 25}            // 8 col offsets
+  // Source: tools/probe_nax_descriptor.py layout-dump variant on g16s.
+  // (Note: spec's dr/dc labels were swapped vs the actual cooperative-tensor layout — corrected in commit at Task 1.6b. The numerical values are unchanged.)
+  METAL_FUNC static short2 get_coord() {
+    const ushort lane = __metal_get_thread_index_in_simdgroup(ushort());
+    const short a = ((lane & 2) >> 1) | ((lane & 4) >> 1) | ((lane & 16) >> 2);  // 8 values — row base
+    const short b = ((lane & 1) << 1) | ((lane & 8) >> 1);                       // 4 values — col base
+    return short2{b, a};  // short2 is {x = col, y = row}
+  }
+
+  METAL_FUNC static short2 dr_dc(short i) {
+    constexpr short dr_table[4] = {0, 8, 16, 24};                  // 4 row offsets
+    constexpr short dc_table[8] = {0, 1, 8, 9, 16, 17, 24, 25};   // 8 col offsets
+    return short2{dc_table[i % 8], dr_table[i / 8]};               // {col_off, row_off}
+  }
+
+  template <
+      typename CType,
+      typename AType,
+      typename BType,
+      bool transpose_a = false,
+      bool transpose_b = false>
+  METAL_FUNC static constexpr void mma(
+      thread dtype_frag_t<CType>& C,
+      const thread dtype_frag_t<AType>& A,
+      metal::bool_constant<transpose_a>,
+      const thread dtype_frag_t<BType>& B,
+      metal::bool_constant<transpose_b>) {
+    constexpr auto desc = mpp::tensor_ops::matmul2d_descriptor(
+        32,
+        32,
+        32,
+        transpose_a,
+        transpose_b,
+        true,
+        mpp::tensor_ops::matmul2d_descriptor::mode::multiply_accumulate);
+
+    mpp::tensor_ops::matmul2d<desc, metal::execution_simdgroup> gemm_op;
+
+    auto ct_a =
+        gemm_op
+            .template get_left_input_cooperative_tensor<AType, BType, CType>();
+    auto ct_b =
+        gemm_op
+            .template get_right_input_cooperative_tensor<AType, BType, CType>();
+    auto ct_c = gemm_op.template get_destination_cooperative_tensor<
+        decltype(ct_a),
+        decltype(ct_b),
+        CType>();
+
+    STEEL_PRAGMA_UNROLL
+    for (short i = 0; i < kElemsPerFrag; i++) {
+      ct_a[i] = A[i];
+      ct_b[i] = B[i];
+      ct_c[i] = C[i];
+    }
+
+    gemm_op.run(ct_a, ct_b, ct_c);
+
+    STEEL_PRAGMA_UNROLL
+    for (short i = 0; i < kElemsPerFrag; i++) {
+      C[i] = ct_c[i];
+    }
+  }
+
+  // Load the entire 32x32 frag from a contiguous threadgroup region.
+  // role and transpose must match the descriptor used by the consumer mma —
+  // MPP's cooperative-tensor per-thread layout depends on (role, transpose).
+  // Materializes a tensor_inline view over the threadgroup pointer and uses
+  // the cooperative-tensor load that the MPP runtime handles. Avoids
+  // hard-coding (dr, dc) per-thread offsets — same code path as
+  // tools/probe_nax_descriptor.py's tg-staged variant, which empirically
+  // verifies max|err| = 0 for the (32, 32, 32) descriptor on g16s.
+  template <Role role, bool transpose, typename T, typename U>
+  METAL_FUNC static void load(
+      thread dtype_frag_t<T>& dst,
+      const threadgroup U* src,
+      const short ld) {
+    constexpr auto desc = mpp::tensor_ops::matmul2d_descriptor(
+        32, 32, 32,
+        /*transpose_left=*/  (role == Role::Left)  ? transpose : false,
+        /*transpose_right=*/ (role == Role::Right) ? transpose : false,
+        /*relaxed_precision=*/true,
+        mpp::tensor_ops::matmul2d_descriptor::mode::multiply_accumulate);
+    mpp::tensor_ops::matmul2d<desc, metal::execution_simdgroup> op;
+
+    metal::dextents<int32_t, 2> ext(32, ld);
+    // MPP cooperative_tensor::load requires the view's element type to match
+    // the cooperative tensor's element type exactly (no const qualifier).
+    // Casting away const internally preserves the const-correct API.
+    metal::tensor<threadgroup U, metal::dextents<int32_t, 2>, metal::tensor_inline>
+        view(const_cast<threadgroup U*>(src), ext);
+
+    // Use if constexpr — Metal may not allow ternary across cooperative_tensor
+    // types, since each branch returns a distinct type.
+    if constexpr (role == Role::Right) {
+      auto ct = op.template get_right_input_cooperative_tensor<T, T, T>();
+      ct.load(view);
+      STEEL_PRAGMA_UNROLL
+      for (short i = 0; i < kElemsPerFrag; i++) dst[i] = static_cast<T>(ct[i]);
+    } else {
+      auto ct = op.template get_left_input_cooperative_tensor<T, T, T>();
+      ct.load(view);
+      STEEL_PRAGMA_UNROLL
+      for (short i = 0; i < kElemsPerFrag; i++) dst[i] = static_cast<T>(ct[i]);
+    }
+  }
+
+  template <typename T, typename U>
+  METAL_FUNC static void store(
+      const thread dtype_frag_t<T>& src,
+      threadgroup U* dst,
+      const short ld) {
+    constexpr auto desc = mpp::tensor_ops::matmul2d_descriptor(
+        32, 32, 32,
+        /*transpose_left=*/false,
+        /*transpose_right=*/false,
+        /*relaxed_precision=*/true,
+        mpp::tensor_ops::matmul2d_descriptor::mode::multiply_accumulate);
+    mpp::tensor_ops::matmul2d<desc, metal::execution_simdgroup> op;
+    auto ct = op.template get_destination_cooperative_tensor<
+        decltype(op.template get_left_input_cooperative_tensor<T, T, T>()),
+        decltype(op.template get_right_input_cooperative_tensor<T, T, T>()),
+        T>();
+
+    STEEL_PRAGMA_UNROLL
+    for (short i = 0; i < kElemsPerFrag; i++) {
+      ct[i] = static_cast<T>(src[i]);
+    }
+
+    metal::dextents<int32_t, 2> ext(32, ld);
+    metal::tensor<threadgroup U, metal::dextents<int32_t, 2>, metal::tensor_inline>
+        view((threadgroup U*)dst, ext);
+    ct.store(view);
+  }
+
+  // Device-memory variant of the contiguous load. role and transpose must
+  // match the consumer mma's descriptor (same reason as the threadgroup
+  // overload). Cooperative tensor is parameterized on U so MPP's
+  // tensor_inline view (also U) accepts ct.load(view). Loaded values are
+  // static_cast<T>(ct[i]) into the thread frag, mirroring the threadgroup
+  // overload's pattern. Validated by
+  // tools/probe_nax_frag32.py::test_mma_via_dv_load_store.
+  template <Role role, bool transpose, typename T, typename U>
+  METAL_FUNC static void load(
+      thread dtype_frag_t<T>& dst,
+      const device U* src,
+      const int ld) {
+    constexpr auto desc = mpp::tensor_ops::matmul2d_descriptor(
+        32, 32, 32,
+        /*transpose_left=*/  (role == Role::Left)  ? transpose : false,
+        /*transpose_right=*/ (role == Role::Right) ? transpose : false,
+        /*relaxed_precision=*/true,
+        mpp::tensor_ops::matmul2d_descriptor::mode::multiply_accumulate);
+    mpp::tensor_ops::matmul2d<desc, metal::execution_simdgroup> op;
+
+    metal::dextents<int32_t, 2> ext(32, ld);
+    metal::tensor<device U, metal::dextents<int32_t, 2>, metal::tensor_inline>
+        view(const_cast<device U*>(src), ext);
+
+    if constexpr (role == Role::Right) {
+      auto ct = op.template get_right_input_cooperative_tensor<U, U, U>();
+      ct.load(view);
+      STEEL_PRAGMA_UNROLL
+      for (short i = 0; i < kElemsPerFrag; i++) dst[i] = static_cast<T>(ct[i]);
+    } else {
+      auto ct = op.template get_left_input_cooperative_tensor<U, U, U>();
+      ct.load(view);
+      STEEL_PRAGMA_UNROLL
+      for (short i = 0; i < kElemsPerFrag; i++) dst[i] = static_cast<T>(ct[i]);
+    }
+  }
+
+  // Device-memory variant of the contiguous store. Same body as the
+  // threadgroup overload with the address space swapped, plus support for
+  // T != U: the destination cooperative tensor is parameterized on the
+  // device pointer type U so MPP's tensor_inline view (also U) accepts
+  // ct.store(view). Frag elements are cast static_cast<U>(src[i]) before
+  // being written into the cooperative tensor.
+  // Validated by tools/probe_nax_frag32.py::test_mma_via_dv_load_store and
+  // ::test_store_device_mixed_precision.
+  template <typename T, typename U>
+  METAL_FUNC static void store(
+      const thread dtype_frag_t<T>& src,
+      device U* dst,
+      const int ld) {
+    constexpr auto desc = mpp::tensor_ops::matmul2d_descriptor(
+        32, 32, 32,
+        /*transpose_left=*/false,
+        /*transpose_right=*/false,
+        /*relaxed_precision=*/true,
+        mpp::tensor_ops::matmul2d_descriptor::mode::multiply_accumulate);
+    mpp::tensor_ops::matmul2d<desc, metal::execution_simdgroup> op;
+    auto ct = op.template get_destination_cooperative_tensor<
+        decltype(op.template get_left_input_cooperative_tensor<U, U, U>()),
+        decltype(op.template get_right_input_cooperative_tensor<U, U, U>()),
+        U>();
+
+    STEEL_PRAGMA_UNROLL
+    for (short i = 0; i < kElemsPerFrag; i++) {
+      ct[i] = static_cast<U>(src[i]);
+    }
+
+    metal::dextents<int32_t, 2> ext(32, ld);
+    metal::tensor<device U, metal::dextents<int32_t, 2>, metal::tensor_inline>
+        view((device U*)dst, ext);
+    ct.store(view);
+  }
+
+  // Load a 32x32 frag from device memory, zero-fill out-of-bounds elements.
+  // Strategy B: caller provides a threadgroup scratch buffer (32*32 elements).
+  // Each lane fills its slice of the staging buffer with bounds-checked copies
+  // from device, then calls the cooperative-tensor load on the staged data.
+  // role and transpose are forwarded to load() so the cooperative-tensor
+  // descriptor matches the consumer mma.
+  // Avoids depending on MPP partial-tile support that does not exist on this
+  // SDK (only tensor_inline/tensor_handle/tensor_offset exist — no padded or
+  // strided bounded-view types); per-thread layout still owned by the
+  // cooperative tensor inside load().
+  // Note: threadgroup scratch cannot be declared inside a device function in
+  // Metal; callers must allocate `threadgroup T scratch[32*32]` and pass it.
+  template <Role role, bool transpose, typename T, typename U>
+  METAL_FUNC static void load_safe(
+      thread dtype_frag_t<T>& dst,
+      const device U* src,
+      const int ld,
+      const short row_lim,
+      const short col_lim,
+      threadgroup T* scratch) {
+    const ushort lane = __metal_get_thread_index_in_simdgroup(ushort());
+    STEEL_PRAGMA_UNROLL
+    for (short i = 0; i < kElemsPerFrag; ++i) {
+      const short flat = lane * kElemsPerFrag + i;
+      const short r = flat / 32;
+      const short c = flat % 32;
+      scratch[r * 32 + c] =
+          (r < row_lim && c < col_lim) ? static_cast<T>(src[r * ld + c]) : T(0);
+    }
+    threadgroup_barrier(mem_flags::mem_threadgroup);
+    load<role, transpose>(dst, scratch, /*ld=*/short(32));
+  }
+
+  // Load a 32x32 frag from device memory with a row bound only; columns are
+  // always treated as full 32. Zero-fills rows at or beyond row_lim.
+  // role and transpose are forwarded to load() so the cooperative-tensor
+  // descriptor matches the consumer mma.
+  // Caller must allocate `threadgroup T scratch[32*32]` and pass it.
+  template <Role role, bool transpose, typename T, typename U>
+  METAL_FUNC static void load_rows(
+      thread dtype_frag_t<T>& dst,
+      const device U* src,
+      const int ld,
+      const short row_lim,
+      threadgroup T* scratch) {
+    const ushort lane = __metal_get_thread_index_in_simdgroup(ushort());
+    STEEL_PRAGMA_UNROLL
+    for (short i = 0; i < kElemsPerFrag; ++i) {
+      const short flat = lane * kElemsPerFrag + i;
+      const short r = flat / 32;
+      const short c = flat % 32;
+      scratch[r * 32 + c] =
+          (r < row_lim) ? static_cast<T>(src[r * ld + c]) : T(0);
+    }
+    threadgroup_barrier(mem_flags::mem_threadgroup);
+    load<role, transpose>(dst, scratch, /*ld=*/short(32));
+  }
+
+  // Store a 32x32 frag to device memory, skipping elements out-of-bounds.
+  // Caller must allocate `threadgroup T scratch[32*32]` and pass it.
+  template <typename T, typename U>
+  METAL_FUNC static void store_safe(
+      const thread dtype_frag_t<T>& src,
+      device U* dst,
+      const int ld,
+      const short row_lim,
+      const short col_lim,
+      threadgroup T* scratch) {
+    store(src, scratch, /*ld=*/short(32));
+    threadgroup_barrier(mem_flags::mem_threadgroup);
+    const ushort lane = __metal_get_thread_index_in_simdgroup(ushort());
+    STEEL_PRAGMA_UNROLL
+    for (short i = 0; i < kElemsPerFrag; ++i) {
+      const short flat = lane * kElemsPerFrag + i;
+      const short r = flat / 32;
+      const short c = flat % 32;
+      if (r < row_lim && c < col_lim) {
+        dst[r * ld + c] = static_cast<U>(scratch[r * 32 + c]);
+      }
+    }
+  }
+
+  // Store a 32x32 frag to device memory with a row bound only; skips rows at
+  // or beyond row_lim. Columns always written for valid rows.
+  // Caller must allocate `threadgroup T scratch[32*32]` and pass it.
+  template <typename T, typename U>
+  METAL_FUNC static void store_rows(
+      const thread dtype_frag_t<T>& src,
+      device U* dst,
+      const int ld,
+      const short row_lim,
+      threadgroup T* scratch) {
+    store(src, scratch, /*ld=*/short(32));
+    threadgroup_barrier(mem_flags::mem_threadgroup);
+    const ushort lane = __metal_get_thread_index_in_simdgroup(ushort());
+    STEEL_PRAGMA_UNROLL
+    for (short i = 0; i < kElemsPerFrag; ++i) {
+      const short flat = lane * kElemsPerFrag + i;
+      const short r = flat / 32;
+      const short c = flat % 32;
+      if (r < row_lim) {
+        dst[r * ld + c] = static_cast<U>(scratch[r * 32 + c]);
+      }
+    }
+  }
+
+  // Row reduction: combine all 32 column elements in each of the thread's
+  // 4 owned rows. Each row is shared by 4 lanes (those with the same
+  // row_base, varying col_base). Lane bits 0 and 3 vary across these 4
+  // lanes; bits 1, 2, 4 are constant. Within each thread, 8 col elements
+  // per row at indices [row*8 .. row*8+7].
+  template <typename Op, typename T>
+  METAL_FUNC static constexpr void row_reduce(
+      thread const dtype_frag_t<T>& inp_vals,
+      thread T* reduced_vals) {
+    STEEL_PRAGMA_UNROLL
+    for (short i = 0; i < kElemRows; i++) {
+      // Within-thread reduction: combine the 8 col elements for this row.
+      T thr_reduce = inp_vals[i * kElemCols + 0];
+      STEEL_PRAGMA_UNROLL
+      for (short j = 1; j < kElemCols; j++) {
+        thr_reduce = Op::apply(thr_reduce, inp_vals[i * kElemCols + j]);
+      }
+
+      // Cross-lane: combine 4 lanes sharing this row_base. Lanes differ
+      // only in bits 0 and 3 of the simd lane id.
+      T shuf1 = simd_shuffle_xor(thr_reduce, ushort(1));
+      thr_reduce = Op::apply(thr_reduce, shuf1);
+      T shuf8 = simd_shuffle_xor(thr_reduce, ushort(8));
+      thr_reduce = Op::apply(thr_reduce, shuf8);
+
+      reduced_vals[i] = Op::apply(reduced_vals[i], thr_reduce);
+    }
+  }
+
+  // Apply a binary op between each row's elements and a per-row scalar.
+  // dtype_frag_t element ordering: row i lives at indices [i*8 .. i*8+7].
+  template <typename Op, typename T>
+  METAL_FUNC static constexpr void row_bin_op(
+      thread dtype_frag_t<T>& inp_vals,
+      thread T* row_vals) {
+    STEEL_PRAGMA_UNROLL
+    for (short i = 0; i < kElemRows; i++) {
+      STEEL_PRAGMA_UNROLL
+      for (short j = 0; j < kElemCols; j++) {
+        inp_vals[i * kElemCols + j] =
+            Op::apply(inp_vals[i * kElemCols + j], row_vals[i]);
+      }
+    }
+  }
+
+  // Store a 32x32 frag to device memory, writing only elements inside the
+  // rectangular slice [start_x, stop_x) x [start_y, stop_y). Out-of-bounds
+  // device positions are not touched. Caller must allocate
+  // `threadgroup T scratch[32*32]` and pass it. The signature mirrors
+  // BaseNAXFrag::store_slice so NAXTile can dispatch with a single call site
+  // (gated on kPacking).
+  template <typename T, typename U, typename StrX, typename StrY,
+            typename StartX, typename StopX, typename StartY, typename StopY,
+            typename OffX = Int<0>, typename OffY = Int<0>>
+  METAL_FUNC static void store_slice(
+      const thread dtype_frag_t<T>& src,
+      device U* dst,
+      StrX str_x,
+      StrY str_y,
+      StartX start_x,
+      StopX stop_x,
+      StartY start_y,
+      StopY stop_y,
+      threadgroup T* scratch,
+      OffX off_x = Int<0>{},
+      OffY off_y = Int<0>{}) {
+    // Stage the 32x32 frag to threadgroup scratch via the contiguous store.
+    store(src, scratch, /*ld=*/short(32));
+    threadgroup_barrier(mem_flags::mem_threadgroup);
+
+    // Per-thread bounded copy: each lane owns kElemsPerFrag (=32) elements.
+    // Address them by flat index: flat = lane * 32 + i, r = flat / 32,
+    // c = flat % 32. The slice bounds are expressed in NAXTile coordinates:
+    // off_x/off_y are the frag's top-left in tile coords (set by NAXTile's
+    // multi-frag dispatch), so a frag-local (r, c) maps to tile-local
+    // (r + off_x, c + off_y).
+    const ushort lane = __metal_get_thread_index_in_simdgroup(ushort());
+    STEEL_PRAGMA_UNROLL
+    for (short i = 0; i < kElemsPerFrag; ++i) {
+      const short flat = lane * kElemsPerFrag + i;
+      const short r = flat / 32;
+      const short c = flat % 32;
+      const short tile_r = r + short(off_x);
+      const short tile_c = c + short(off_y);
+      if (tile_r >= short(start_x) && tile_r < short(stop_x) &&
+          tile_c >= short(start_y) && tile_c < short(stop_y)) {
+        dst[r * str_x + c * str_y] = static_cast<U>(scratch[r * 32 + c]);
+      }
+    }
+  }
+};
+
+template <
+    typename T,
+    short kTileRows_,
+    short kTileCols_,
+    class NAXFrag_ = BaseNAXFrag>
+struct NAXTile {
+  using NAXFrag_t = NAXFrag_;
+  using elem_type = T;
+
+  STEEL_CONST short kFragRows = NAXFrag_t::kFragRows;
+  STEEL_CONST short kFragCols = NAXFrag_t::kFragCols;
+  STEEL_CONST short kElemsPerFrag = NAXFrag_t::kElemsPerFrag;
+
+  STEEL_CONST short kTileRows = kTileRows_;
+  STEEL_CONST short kTileCols = kTileCols_;
+
+  STEEL_CONST short kRows = kTileRows * kFragRows;
+  STEEL_CONST short kCols = kTileCols * kFragCols;
+
+  STEEL_CONST short kNumFrags = kTileRows * kTileCols;
+  STEEL_CONST short kElemsPerTile = kNumFrags * kElemsPerFrag;
+
+  STEEL_CONST short kFragThrRows = NAXFrag_t::kElemRows;
+  STEEL_CONST short kFragThrCols = NAXFrag_t::kElemCols;
+  STEEL_CONST short kFragRowsJump = NAXFrag_t::kElemRowsJump;
+
+  STEEL_CONST short kRowsPerThread = kTileRows * NAXFrag_t::kElemRows;
+  STEEL_CONST short kColsPerThread = kTileCols * NAXFrag_t::kElemCols;
+
+  typedef typename NAXFrag_t::template dtype_frag_t<T> frag_type;
+
+  frag_type val_frags[kNumFrags]; // = {frag_type(0)};
+
+  METAL_FUNC NAXTile() thread {}
+
+  METAL_FUNC constexpr void clear() {
+    STEEL_PRAGMA_UNROLL
+    for (short i = 0; i < kNumFrags; ++i) {
+      val_frags[i] = frag_type(0);
+    }
+  }
+
+  METAL_FUNC constexpr thread frag_type& frag_at(const short i, const short j) {
+    return val_frags[i * kTileCols + j];
+  }
+
+  METAL_FUNC constexpr const thread frag_type& frag_at(
+      const short i,
+      const short j) const {
+    return val_frags[i * kTileCols + j];
+  }
+
+  template <int i, int j>
+  METAL_FUNC constexpr thread frag_type& frag_at() {
+    return val_frags[i * kTileCols + j];
+  }
+
+  template <int i, int j>
+  METAL_FUNC constexpr const thread frag_type& frag_at() const {
+    return val_frags[i * kTileCols + j];
+  }
+
+  template <bool transpose>
+  METAL_FUNC constexpr thread frag_type&
+  frag_at(const short i, const short j, metal::bool_constant<transpose>) {
+    if constexpr (transpose) {
+      return frag_at(j, i);
+    } else {
+      return frag_at(i, j);
+    }
+  }
+
+  template <bool transpose>
+  METAL_FUNC constexpr const thread frag_type&
+  frag_at(const short i, const short j, metal::bool_constant<transpose>) const {
+    if constexpr (transpose) {
+      return frag_at(j, i);
+    } else {
+      return frag_at(i, j);
+    }
+  }
+
+  template <int i, int j, bool transpose>
+  METAL_FUNC constexpr thread frag_type& frag_at() {
+    if constexpr (transpose) {
+      return frag_at<j, i>();
+    } else {
+      return frag_at<i, j>();
+    }
+  }
+
+  template <int i, int j, bool transpose>
+  METAL_FUNC constexpr const thread frag_type& frag_at() const {
+    if constexpr (transpose) {
+      return frag_at<j, i>();
+    } else {
+      return frag_at<i, j>();
+    }
+  }
+
+  METAL_FUNC thread elem_type* elems() {
+    return reinterpret_cast<thread elem_type*>(val_frags);
+  }
+
+  METAL_FUNC const thread elem_type* elems() const {
+    return reinterpret_cast<const thread elem_type*>(val_frags);
+  }
+
+  template <typename Op>
+  METAL_FUNC void row_reduce(thread metal::vec<T, kRowsPerThread>& vals) const {
+    auto vptr = (thread T*)(&vals);
+    STEEL_PRAGMA_UNROLL
+    for (short i = 0; i < kTileRows; ++i) {
+      STEEL_PRAGMA_UNROLL
+      for (short j = 0; j < kTileCols; ++j) {
+        NAXFrag_t::template row_reduce<Op>(
+            frag_at(i, j), &vptr[i * kFragThrRows]);
+      }
+    }
+  }
+
+  template <typename Op>
+  METAL_FUNC void row_bin_op(thread metal::vec<T, kRowsPerThread>& vals) {
+    auto vptr = (thread T*)(&vals);
+    STEEL_PRAGMA_UNROLL
+    for (short i = 0; i < kTileRows; ++i) {
+      STEEL_PRAGMA_UNROLL
+      for (short j = 0; j < kTileCols; ++j) {
+        NAXFrag_t::template row_bin_op<Op>(
+            frag_at(i, j), &vptr[i * kFragThrRows]);
+      }
+    }
+  }
+
+  template <typename U, int str_x, int str_y>
+  METAL_FUNC void load(const threadgroup U* src) {
+    const_for_loop<0, kTileRows, 1>([&](auto idx_row) {
+      const_for_loop<0, kTileCols, 1>([&](auto idx_col) {
+        NAXFrag_t::load(
+            frag_at<idx_row.value, idx_col.value>(),
+            src,
+            Int<str_x>{},
+            Int<str_y>{},
+            idx_row * Int<kFragRows>{},
+            idx_col * Int<kFragCols>{});
+      });
+    });
+  }
+
+  template <typename U, int str_x, int str_y>
+  METAL_FUNC void store(threadgroup U* dst) const {
+    const_for_loop<0, kTileRows, 1>([&](auto idx_row) {
+      const_for_loop<0, kTileCols, 1>([&](auto idx_col) {
+        NAXFrag_t::store(
+            frag_at<idx_row.value, idx_col.value>(),
+            dst,
+            Int<str_x>{},
+            Int<str_y>{},
+            idx_row * Int<kFragRows>{},
+            idx_col * Int<kFragCols>{});
+      });
+    });
+  }
+
+  // role and transpose are only meaningful for NAXFrag32 (kPacking==1).
+  // For BaseNAXFrag (kPacking==2) the params are ignored. No defaults are
+  // provided — callers must pass explicit values so a future kPacking==1
+  // caller cannot silently pick wrong params and reintroduce the g16 bug.
+  template <Role role, bool transpose, typename U>
+  METAL_FUNC void load(const device U* src, const int ld,
+                       threadgroup elem_type* scratch) {
+    if constexpr (NAXFrag_t::kPacking == 1) {
+      // kPacking==1 path: NAXFrag32's cooperative-tensor device load only
+      // works correctly when ld == kFragCols (==32). For any other stride,
+      // stage through threadgroup scratch (same as load_rows with full rows).
+      const_for_loop<0, kTileRows, 1>([&](auto idx_row) {
+        const_for_loop<0, kTileCols, 1>([&](auto idx_col) {
+          constexpr short m_off = idx_row.value * kFragRows;
+          constexpr short n_off = idx_col.value * kFragCols;
+          NAXFrag_t::template load_rows<role, transpose>(
+              frag_at<idx_row.value, idx_col.value>(),
+              src + m_off * ld + n_off,
+              ld,
+              short(kFragRows),
+              scratch);
+        });
+      });
+    } else {
+      // kPacking==2 (BaseNAXFrag): role/transpose not forwarded — per-element
+      // dr/dc loads are layout-invariant.
+      const_for_loop<0, kTileRows, 1>([&](auto idx_row) {
+        const_for_loop<0, kTileCols, 1>([&](auto idx_col) {
+          NAXFrag_t::load(
+              frag_at<idx_row.value, idx_col.value>(),
+              src,
+              ld,
+              Int<1>{},
+              idx_row * Int<kFragRows>{},
+              idx_col * Int<kFragCols>{});
+        });
+      });
+    }
+  }
+
+  template <typename U>
+  METAL_FUNC void store(device U* dst, const int ld,
+                        threadgroup elem_type* scratch) const {
+    if constexpr (NAXFrag_t::kPacking == 1) {
+      // kPacking==1 path: NAXFrag32's cooperative-tensor device store only
+      // works correctly when ld == kFragCols (==32). For any other stride,
+      // stage through threadgroup scratch (same as store_rows with full rows).
+      const_for_loop<0, kTileRows, 1>([&](auto idx_row) {
+        const_for_loop<0, kTileCols, 1>([&](auto idx_col) {
+          constexpr short m_off = idx_row.value * kFragRows;
+          constexpr short n_off = idx_col.value * kFragCols;
+          NAXFrag_t::store_rows(
+              frag_at<idx_row.value, idx_col.value>(),
+              dst + m_off * ld + n_off,
+              ld,
+              short(kFragRows),
+              scratch);
+        });
+      });
+    } else {
+      const_for_loop<0, kTileRows, 1>([&](auto idx_row) {
+        const_for_loop<0, kTileCols, 1>([&](auto idx_col) {
+          NAXFrag_t::store(
+              frag_at<idx_row.value, idx_col.value>(),
+              dst,
+              ld,
+              Int<1>{},
+              idx_row * Int<kFragRows>{},
+              idx_col * Int<kFragCols>{});
+        });
+      });
+    }
+  }
+
+  // Convenience overloads for BaseNAXFrag (kPacking==2) callers that do not
+  // need role/transpose or scratch. These match the old attn/nax.h NAXTile API
+  // and are only valid when NAXFrag_t::kPacking == 2.  Calling them with a
+  // kPacking==1 frag is a compile-time error (static_assert).
+
+  template <typename U>
+  METAL_FUNC void load(const device U* src, const int ld) {
+    static_assert(
+        NAXFrag_t::kPacking == 2,
+        "NAXTile::load(src, ld) is only valid for kPacking==2 (BaseNAXFrag); "
+        "use load<role, transpose>(src, ld, scratch) for NAXFrag32");
+    const_for_loop<0, kTileRows, 1>([&](auto idx_row) {
+      const_for_loop<0, kTileCols, 1>([&](auto idx_col) {
+        NAXFrag_t::load(
+            frag_at<idx_row.value, idx_col.value>(),
+            src,
+            ld,
+            Int<1>{},
+            idx_row * Int<kFragRows>{},
+            idx_col * Int<kFragCols>{});
+      });
+    });
+  }
+
+  template <typename U>
+  METAL_FUNC void load_rows(const device U* src, const int ld,
+                            const short n_rows) {
+    static_assert(
+        NAXFrag_t::kPacking == 2,
+        "NAXTile::load_rows(src, ld, n_rows) is only valid for kPacking==2 "
+        "(BaseNAXFrag); use load_rows<role, transpose>(src, ld, n_rows, "
+        "scratch) for NAXFrag32");
+    const_for_loop<0, kTileRows, 1>([&](auto idx_row) {
+      const_for_loop<0, kTileCols, 1>([&](auto idx_col) {
+        NAXFrag_t::load_rows(
+            frag_at<idx_row.value, idx_col.value>(),
+            src,
+            ld,
+            Int<1>{},
+            n_rows,
+            idx_row * Int<kFragRows>{},
+            idx_col * Int<kFragCols>{});
+      });
+    });
+  }
+
+  template <typename U>
+  METAL_FUNC void store(device U* dst, const int ld) const {
+    static_assert(
+        NAXFrag_t::kPacking == 2,
+        "NAXTile::store(dst, ld) is only valid for kPacking==2 (BaseNAXFrag); "
+        "use store(dst, ld, scratch) for NAXFrag32");
+    const_for_loop<0, kTileRows, 1>([&](auto idx_row) {
+      const_for_loop<0, kTileCols, 1>([&](auto idx_col) {
+        NAXFrag_t::store(
+            frag_at<idx_row.value, idx_col.value>(),
+            dst,
+            ld,
+            Int<1>{},
+            idx_row * Int<kFragRows>{},
+            idx_col * Int<kFragCols>{});
+      });
+    });
+  }
+
+  // role and transpose must be specified explicitly — no defaults, so a future
+  // kPacking==1 caller cannot silently pick wrong params.
+  template <Role role, bool transpose, typename U>
+  METAL_FUNC void
+  load_rows(const device U* src, const int ld, const short n_rows,
+            threadgroup elem_type* scratch = nullptr) {
+    if constexpr (NAXFrag_t::kPacking == 1) {
+      // kPacking==1 path: stages via NAXFrag32::load_rows scratch machinery.
+      const_for_loop<0, kTileRows, 1>([&](auto idx_row) {
+        const_for_loop<0, kTileCols, 1>([&](auto idx_col) {
+          constexpr short m_off = idx_row.value * kFragRows;
+          constexpr short n_off = idx_col.value * kFragCols;
+          // Per-frag row bound clamped to [0, kFragRows].
+          const short rl = max(short(0),
+                               min(short(kFragRows), short(n_rows - m_off)));
+          NAXFrag_t::template load_rows<role, transpose>(
+              frag_at<idx_row.value, idx_col.value>(),
+              src + m_off * ld + n_off,
+              ld,
+              rl,
+              scratch);
+        });
+      });
+    } else {
+      // kPacking==2 (BaseNAXFrag) path: unchanged; scratch param ignored.
+      const_for_loop<0, kTileRows, 1>([&](auto idx_row) {
+        const_for_loop<0, kTileCols, 1>([&](auto idx_col) {
+          NAXFrag_t::load_rows(
+              frag_at<idx_row.value, idx_col.value>(),
+              src,
+              ld,
+              Int<1>{},
+              n_rows,
+              idx_row * Int<kFragRows>{},
+              idx_col * Int<kFragCols>{});
+        });
+      });
+    }
+  }
+
+  // role and transpose must be specified explicitly — no defaults, so a future
+  // kPacking==1 caller cannot silently pick wrong params.
+  template <Role role, bool transpose, typename U>
+  METAL_FUNC void
+  load_safe(const device U* src, const int ld, const short2 src_tile_dims,
+            threadgroup elem_type* scratch = nullptr) {
+    if constexpr (NAXFrag_t::kPacking == 1) {
+      // kPacking==1 path: stages via NAXFrag32::load_safe scratch machinery.
+      const_for_loop<0, kTileRows, 1>([&](auto idx_row) {
+        const_for_loop<0, kTileCols, 1>([&](auto idx_col) {
+          constexpr short m_off = idx_row.value * kFragRows;
+          constexpr short n_off = idx_col.value * kFragCols;
+          // Per-frag row/col bounds clamped to [0, kFragRows/kFragCols].
+          const short rl = max(short(0),
+                               min(short(kFragRows),
+                                   short(src_tile_dims.y - m_off)));
+          const short cl = max(short(0),
+                               min(short(kFragCols),
+                                   short(src_tile_dims.x - n_off)));
+          NAXFrag_t::template load_safe<role, transpose>(
+              frag_at<idx_row.value, idx_col.value>(),
+              src + m_off * ld + n_off,
+              ld,
+              rl,
+              cl,
+              scratch);
+        });
+      });
+    } else {
+      // kPacking==2 (BaseNAXFrag) path: unchanged; scratch param ignored.
+      const_for_loop<0, kTileRows, 1>([&](auto idx_row) {
+        const_for_loop<0, kTileCols, 1>([&](auto idx_col) {
+          NAXFrag_t::load_safe(
+              frag_at<idx_row.value, idx_col.value>(),
+              src,
+              ld,
+              Int<1>{},
+              src_tile_dims.y,
+              src_tile_dims.x,
+              idx_row * Int<kFragRows>{},
+              idx_col * Int<kFragCols>{});
+        });
+      });
+    }
+  }
+
+  template <typename U>
+  METAL_FUNC void store_rows(device U* dst, const int ld, const short n_rows,
+                             threadgroup elem_type* scratch = nullptr) const {
+    if constexpr (NAXFrag_t::kPacking == 1) {
+      // kPacking==1 path: stages via NAXFrag32::store_rows scratch machinery.
+      const_for_loop<0, kTileRows, 1>([&](auto idx_row) {
+        const_for_loop<0, kTileCols, 1>([&](auto idx_col) {
+          constexpr short m_off = idx_row.value * kFragRows;
+          constexpr short n_off = idx_col.value * kFragCols;
+          // Per-frag row bound clamped to [0, kFragRows].
+          const short rl = max(short(0),
+                               min(short(kFragRows), short(n_rows - m_off)));
+          NAXFrag_t::store_rows(
+              frag_at<idx_row.value, idx_col.value>(),
+              dst + m_off * ld + n_off,
+              ld,
+              rl,
+              scratch);
+        });
+      });
+    } else {
+      // kPacking==2 (BaseNAXFrag) path: unchanged; scratch param ignored.
+      const_for_loop<0, kTileRows, 1>([&](auto idx_row) {
+        const_for_loop<0, kTileCols, 1>([&](auto idx_col) {
+          NAXFrag_t::store_rows(
+              frag_at<idx_row.value, idx_col.value>(),
+              dst,
+              ld,
+              Int<1>{},
+              n_rows,
+              idx_row * Int<kFragRows>{},
+              idx_col * Int<kFragCols>{});
+        });
+      });
+    }
+  }
+
+  template <typename U>
+  METAL_FUNC void
+  store_safe(device U* dst, const int ld, const short2 dst_tile_dims,
+             threadgroup elem_type* scratch = nullptr) const {
+    if constexpr (NAXFrag_t::kPacking == 1) {
+      // kPacking==1 path: stages via NAXFrag32::store_safe scratch machinery.
+      const_for_loop<0, kTileRows, 1>([&](auto idx_row) {
+        const_for_loop<0, kTileCols, 1>([&](auto idx_col) {
+          constexpr short m_off = idx_row.value * kFragRows;
+          constexpr short n_off = idx_col.value * kFragCols;
+          // Per-frag row/col bounds clamped to [0, kFragRows/kFragCols].
+          const short rl = max(short(0),
+                               min(short(kFragRows),
+                                   short(dst_tile_dims.y - m_off)));
+          const short cl = max(short(0),
+                               min(short(kFragCols),
+                                   short(dst_tile_dims.x - n_off)));
+          NAXFrag_t::store_safe(
+              frag_at<idx_row.value, idx_col.value>(),
+              dst + m_off * ld + n_off,
+              ld,
+              rl,
+              cl,
+              scratch);
+        });
+      });
+    } else {
+      // kPacking==2 (BaseNAXFrag) path: unchanged; scratch param ignored.
+      const_for_loop<0, kTileRows, 1>([&](auto idx_row) {
+        const_for_loop<0, kTileCols, 1>([&](auto idx_col) {
+          NAXFrag_t::store_safe(
+              frag_at<idx_row.value, idx_col.value>(),
+              dst,
+              ld,
+              Int<1>{},
+              dst_tile_dims.y,
+              dst_tile_dims.x,
+              idx_row * Int<kFragRows>{},
+              idx_col * Int<kFragCols>{});
+        });
+      });
+    }
+  }
+
+  template <typename U>
+  METAL_FUNC void store_slice(
+      device U* dst,
+      const int ld,
+      const short2 start,
+      const short2 stop,
+      threadgroup elem_type* scratch = nullptr) const {
+    const_for_loop<0, kTileRows, 1>([&](auto idx_row) {
+      const_for_loop<0, kTileCols, 1>([&](auto idx_col) {
+        if constexpr (NAXFrag_t::kPacking == 1) {
+          // NAXFrag32 (g16) — pass scratch.
+          NAXFrag_t::store_slice(
+              frag_at<idx_row.value, idx_col.value>(),
+              dst,
+              ld,
+              Int<1>{},
+              start.y, stop.y,
+              start.x, stop.x,
+              scratch,
+              idx_row * Int<kFragRows>{},
+              idx_col * Int<kFragCols>{});
+        } else {
+          // BaseNAXFrag — no scratch arg.
+          NAXFrag_t::store_slice(
+              frag_at<idx_row.value, idx_col.value>(),
+              dst,
+              ld,
+              Int<1>{},
+              start.y, stop.y,
+              start.x, stop.x,
+              idx_row * Int<kFragRows>{},
+              idx_col * Int<kFragCols>{});
+        }
+      });
+    });
+  }
+};
+
+template <
+    class CTile,
+    class ATile,
+    class BTile,
+    bool transpose_a,
+    bool transpose_b>
+METAL_FUNC void tile_matmad_nax(
+    thread CTile& C,
+    thread ATile& A,
+    metal::bool_constant<transpose_a>,
+    thread BTile& B,
+    metal::bool_constant<transpose_b>) {
+  // Static checks
+  constexpr short TMa = transpose_a ? ATile::kTileCols : ATile::kTileRows;
+  constexpr short TM = CTile::kTileRows;
+  static_assert(TMa == TM, "MXU tile matmul: M dimensions do not match");
+
+  constexpr short TNb = transpose_b ? BTile::kTileRows : BTile::kTileCols;
+  constexpr short TN = CTile::kTileCols;
+  static_assert(TNb == TN, "MXU tile matmul: N dimensions do not match");
+
+  constexpr short TKa = transpose_a ? ATile::kTileRows : ATile::kTileCols;
+  constexpr short TK = transpose_b ? BTile::kTileCols : BTile::kTileRows;
+  static_assert(TKa == TK, "MXU tile matmul: K dimensions do not match");
+
+  constexpr auto ta = metal::bool_constant<transpose_a>{};
+  constexpr auto tb = metal::bool_constant<transpose_b>{};
+
+  if constexpr (CTile::NAXFrag_t::kPacking == 1) {
+    // Single-frag mma: one 32x32 output per MMA. Used by NAXFrag32 (g16 path).
+    STEEL_PRAGMA_UNROLL
+    for (short mm = 0; mm < TM; ++mm) {
+      STEEL_PRAGMA_UNROLL
+      for (short nn = 0; nn < TN; ++nn) {
+        STEEL_PRAGMA_UNROLL
+        for (short kk = 0; kk < TK; ++kk) {
+          CTile::NAXFrag_t::mma(
+              C.frag_at(mm, nn),
+              A.frag_at(mm, kk, ta),
+              metal::bool_constant<transpose_a>{},
+              B.frag_at(kk, nn, tb),
+              metal::bool_constant<transpose_b>{});
+        }
+      }
+    }
+  } else if constexpr (TN == 1 && TM % 2 == 0) {
+    STEEL_PRAGMA_UNROLL
+    for (short mm = 0; mm < TM; mm += 2) {
+      STEEL_PRAGMA_UNROLL
+      for (short nn = 0; nn < TN; ++nn) {
+        STEEL_PRAGMA_UNROLL
+        for (short kk = 0; kk < TK; ++kk) {
+          CTile::NAXFrag_t::mma(
+              C.frag_at(mm, nn),
+              C.frag_at(mm + 1, nn),
+              A.frag_at(mm, kk, ta),
+              A.frag_at(mm + 1, kk, ta),
+              metal::bool_constant<transpose_a>{},
+              B.frag_at(kk, nn, tb),
+              metal::bool_constant<transpose_b>{});
+        }
+      }
+    }
+  } else if constexpr (TN % 2 == 0) {
+    STEEL_PRAGMA_UNROLL
+    for (short mm = 0; mm < TM; ++mm) {
+      STEEL_PRAGMA_UNROLL
+      for (short nn = 0; nn < TN; nn += 2) {
+        STEEL_PRAGMA_UNROLL
+        for (short kk = 0; kk < TK; ++kk) {
+          CTile::NAXFrag_t::mma(
+              C.frag_at(mm, nn),
+              C.frag_at(mm, nn + 1),
+              A.frag_at(mm, kk, ta),
+              metal::bool_constant<transpose_a>{},
+              B.frag_at(kk, nn, tb),
+              B.frag_at(kk, nn + 1, tb),
+              metal::bool_constant<transpose_b>{});
+        }
+      }
+    }
+  }
+}
+
+} // namespace steel
+} // namespace mlx
